task-1-test-covering: Class TeamMember test added, Class Tester test refactored
