+++ src/test/java/ru/sberbank/data/TesterTest.java
@@ -31,55 +31,4 @@ class TesterTest {
         assertEquals(task, tester.checkTask());
     }
 
-    @Test
-    void getCurrentTaskTest_Fail() {    // actually, this is redundant. The method is tested above
-        assertNull(tester.getCurrentTask());
-    }
-
-
-    @Test
-    void addTaskTest_Fail() {
-        String expected = "Нарушение процесса!";
-        tester.addTask(task);
-        Exception exception = assertThrows(IllegalStateException.class,
-                () -> tester.addTask(task));
-        assertEquals(expected, exception.getMessage());
-    }
-
-
-    @Test
-    void getFirstNameTest_Success() {
-        String expected = "John";
-        assertEquals(expected, tester.getFirstName());
-    }
-
-    @Test
-    void getSecondNameTest_Success() {
-        String expected = "Doe";
-        assertEquals(expected, tester.getSecondName());
-
-    }
-
-    @Test
-    void isFreeTest_Default() {
-        boolean expected = false;   // default value must be returned
-        assertEquals(expected, tester.isFree());
-    }
-
-    @Test
-    void isFreeTestBusy_Default() {
-        boolean expected = false;
-        tester.addTask(task);
-        assertEquals(expected, tester.isFree());
-    }
-
-   @Test
-    void isFreeTestFree_Success() {
-        boolean expected = true;
-        tester.addTask(task);
-        tester.checkTask();
-        assertEquals(expected, tester.isFree());
-    }
-
-
 }
